+++ src/components/TextField/TextField.tsx
@@ -10,7 +10,7 @@ import {
 import Icon from "../Icon/Icon";
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 import {ValidIconName} from "../Icon/canonicalIconNames";
-import {useTextValidation} from "./useTextValidation";
+import {InvisibleCharacterWarningProps, useTextValidation} from "./useTextValidation";
 
 export interface TextFieldProps extends Partial<Omit<InputGroupProps, "intent" | "leftIcon"> & HTMLInputProps> {
     /**
@@ -38,9 +38,9 @@ export interface TextFieldProps extends Partial<Omit<InputGroupProps, "intent" |
      */
     leftIcon?: ValidIconName | MaybeElement;
     /**
-     * If set, the function is called if any invisible, hard to spot characters in the string value are detected.
+     * If set, allows to be informed of invisible, hard to spot characters in the string value.
      */
-    invisibleCharacterWarningCallback?: (detectedCodePoints: Set<number>) => any
+    invisibleCharacterWarning?: InvisibleCharacterWarningProps
 }
 
 /**
@@ -54,7 +54,7 @@ function TextField({
   hasStateDanger = false,
   fullWidth = true,
   leftIcon,
-  invisibleCharacterWarningCallback,
+  invisibleCharacterWarning,
   ...otherProps
 }: TextFieldProps) {
   let intent;
@@ -75,7 +75,7 @@ function TextField({
       break;
   }
 
-  const maybeWrappedOnChange = useTextValidation({...otherProps, invisibleCharacterWarningCallback})
+  const maybeWrappedOnChange = useTextValidation({...otherProps, invisibleCharacterWarning})
 
   if ((otherProps.readOnly || otherProps.disabled) && !!otherProps.value && !otherProps.title) {
       otherProps["title"] = otherProps.value;